#300: updated resolver concept to introduce an intermediate DefaultOrganizationNameKeyResolver - this should make it easier to override only that logic if the default heuristics aren't sufficient.
