+++ extensions/servlet/src/main/java/com/stormpath/sdk/servlet/idsite/DefaultIdSiteOrganizationResolver.java
@@ -4,21 +4,20 @@ import com.stormpath.sdk.servlet.http.Resolver;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.List;
 
 /**
  * @since 1.0.RC5
  */
 public class DefaultIdSiteOrganizationResolver implements Resolver<IdSiteOrganizationContext> {
 
-    private Resolver<List<String>> subdomainResolver;
+    private Resolver<String> organizationNameKeyResolver;
 
     private Boolean useSubdomain;
 
     private Boolean showOrganizationField;
 
-    public void setSubdomainResolver(Resolver<List<String>> subdomainResolver) {
-        this.subdomainResolver = subdomainResolver;
+    public void setOrganizationNameKeyResolver(Resolver<String> organizationNameKeyResolver) {
+        this.organizationNameKeyResolver = organizationNameKeyResolver;
     }
 
     public void setUseSubdomain(Boolean useSubdomain) {
@@ -32,13 +31,7 @@ public class DefaultIdSiteOrganizationResolver implements Resolver<IdSiteOrganiz
     @Override
     public IdSiteOrganizationContext get(HttpServletRequest request, HttpServletResponse response) {
 
-        List<String> subdomains = subdomainResolver.get(request, null);
-
-        String subdomain = null;
-
-        if (subdomains.size() == 1) {
-            subdomain = subdomains.get(0);
-        }
+        String subdomain = organizationNameKeyResolver.get(request, response);
 
         return new DefaultIdSiteOrganizationContext(subdomain, useSubdomain, showOrganizationField);
     }
