+++ extensions/servlet/src/main/resources/META-INF/com/stormpath/sdk/servlet/default.stormpath.properties
@@ -62,7 +62,9 @@ stormpath.web.idSite.enabled = false
 stormpath.web.idSite.register.uri = /#/register
 stormpath.web.idSite.forgot.uri = /#/forgot
 stormpath.web.idSite.result.uri = /idSiteResult
+# empty property means use ID site defaults:
 stormpath.web.idSite.showOrganizationField =
+# empty property means use ID site defaults:
 stormpath.web.idSite.useSubdomain =
 # Inferred based on heuristics by default.  However if your application is not deployed to an apex domain, like
 # myapp.com, you *must* specify your application's base domain, e.g. myapp.mycompany.com
